refactor: change CryptoUtils to CryptoService
